feat(crypto): Refactor certificate marshaling and add LDAP unmarshaling
This commit refactors the certificate marshaling logic in the mod_crypto package.
The MarshalJSON and UnmarshalJSON methods for Certificate are moved to a new file, marshal_json.go, and now handle P12 encoding/decoding.
A new Certificates type is introduced to represent a map of certificates.
Additionally, a new UnmarshalLDAPAttr method is added to Certificates in marshal_ldap.go to handle unmarshaling from LDAP attributes.
The attrUserPKCS12s type in mod_db is updated to use the new mod_crypto.Certificates type, and its custom JSON and LDAP unmarshaling methods are removed as they are now handled by the mod_crypto package.
This change centralizes certificate marshaling and unmarshaling logic within the mod_crypto package and simplifies the mod_db package.