feat: Update Certificate struct to include P12 and CertificatesPEM fields
This commit adds a `P12` field to the `Certificate` struct to store the P12 file. It also updates the `CertificatesDER` and `CertificatesPEM` fields to be pointers to `[]byte` to accommodate the new `P12` field.  The JSON tags are updated to reflect these changes.  The `Certificate` struct now also includes a `Certificate` field of type `x509.Certificate` and `CertificateCAChain` to store the certificate and the certificate chain.